// src/libs/core-api/api/model/user-search-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BlockInfrastructure-Core
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface UserSearchResponse
 */
export interface UserSearchResponse {
  /**
   *
   * @type {string}
   * @memberof UserSearchResponse
   */
  userId: string;
  /**
   *
   * @type {string}
   * @memberof UserSearchResponse
   */
  userName: string;
  /**
   *
   * @type {string}
   * @memberof UserSearchResponse
   */
  userEmail: string;
  /**
   *
   * @type {string}
   * @memberof UserSearchResponse
   */
  profilePictureImageUrl?: string | null;
}